Self Review: move form component to app folder
